feat: unified startup — serve dashboard from command center
Merge the dashboard into the command center so a single daemon process
serves both the REST API and the pre-built Next.js static export on one
port. This eliminates the separate `pnpm dev` dashboard process that
only worked from a source checkout and never shipped with npm installs.

Key changes:

- Convert dashboard to Next.js static export (output: "export"), split
  dynamic route into Server + Client components for generateStaticParams
- Add Hono static file middleware (src/command-center/static.ts) with
  SPA fallback for /project/* routes and proper cache headers
- Remove startDashboard() / stopDashboard() — no separate process
- Add daemon health-check gate after spawn (waitForDaemon)
- Add service recovery on re-attach (restart dead daemons)
- Multi-session port sharing: when port is already bound by another
  tmux-ide session, store the shared port instead of failing silently
- Fix flaky orchestrator tests: isAtAgentPrompt() now uses the mockable
  captureLastLine() instead of calling execFileSync("tmux") directly
- Ship dashboard/out in npm package via prepublishOnly build step
