+++ dashboard/.gitignore
@@ -1,3 +1,4 @@
 .next
+out/
 node_modules
 next-env.d.ts
