+++ dashboard/next.config.mjs
@@ -5,6 +5,8 @@ const dashboardDir = dirname(fileURLToPath(import.meta.url));
 
 /** @type {import('next').NextConfig} */
 const config = {
+  output: "export",
+  trailingSlash: true,
   reactStrictMode: true,
   turbopack: {
     root: resolve(dashboardDir, ".."),
