+++ src/stop.ts
@@ -1,25 +1,7 @@
 import { resolve } from "node:path";
 import { getSessionName } from "./lib/yaml-io.ts";
 import { outputError } from "./lib/output.ts";
-import { getSessionVariable, killSession, stopSessionMonitor } from "./lib/tmux.ts";
-
-export function stopDashboard(session: string): void {
-  const pid = getSessionVariable(session, "@dashboard_pid");
-  if (!pid) return;
-
-  const numPid = Number.parseInt(pid, 10);
-  if (Number.isNaN(numPid)) return;
-
-  try {
-    process.kill(-numPid, "SIGTERM");
-  } catch {
-    try {
-      process.kill(numPid, "SIGTERM");
-    } catch {
-      // Process already gone.
-    }
-  }
-}
+import { killSession, stopSessionMonitor } from "./lib/tmux.ts";
 
 export async function stop(
   targetDir: string | undefined,
@@ -30,7 +12,6 @@ export async function stop(
 
   // Stop the session monitor before killing the session
   stopSessionMonitor(session);
-  stopDashboard(session);
 
   const result = killSession(session);
